+++ core/remote/src/main/java/kr/co/remote/model/response/GetAccountBookListResponse.kt
@@ -37,7 +37,7 @@ internal data class GetAccountBookListResponse(
 
         @Serializable
         data class PercentCategory(
-            val percent: Int,
+            val percent: Float,
             val category: String
         )
     }
